[server] refactor: unify file naming convention (#30)
