+++ server/includes/sensor.h
